requirements.txt
playwright
pandas
PyPDF2
google-auth
google-auth-oauthlib
google-api-python-client